requirements.txt
absl-py==2.0.0
accelerate==0.23.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall
backports.functools-lru-cache
brotlipy==0.7.0
bytecode==0.15.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
colorama==0.4.6
comm
contourpy==1.0.7
cycler==0.11.0
datasets==2.10.1
debugpy
decorator
diffusers==0.21.2
dill==0.3.6
docker-pycreds==0.4.0
dpcpp-cpp-rt==2023.2.0
easydict==1.11
einops==0.6.0
entrypoints 
exceptiongroup
executin
filelock==3.10.0
fonttools==4.39.2
frozenlist==1.3.3
fsspec==2023.3.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.23.1
google-auth-oauthlib==1.0.0
grpcio==1.58.0
huggingface-hub==0.13.2
idna==3.4
imageio==2.31.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
intel-cmplr-lib-rt==2023.2.0
intel-cmplr-lic-rt==2023.2.0
intel-opencl-rt==2023.2.0
intel-openmp==2023.2.0
ipdb==0.13.13
ipykernel
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
jupyter-client
jupyter_core
kiwisolver==1.4.4
kornia==0.7.0
lazy_loader==0.3
lit==15.0.7
llvmlite==0.41.0
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline
mdurl==0.1.2
mkl==2023.2.0
mkl-fft==1.3.1
mkl-service==2.4.0
mmcv==1.7.1
model-index==0.1.11
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
nest-asyncio
networkx==3.0
ninja==1.11.1.1
numba==0.58.0
numpy==1.22.4
nvidia-ml-py==12.535.108
nvitop==1.3.0
oauthlib==3.2.2
opencv-python==4.7.0.72
openmim==0.3.6
ordered-set==4.1.0
packaging==23.0
pandas==1.5.3
parso
pathtools==0.1.2
pexpect
pickleshare
Pillow==9.4.0
platformdirs
prettytable==3.6.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil
ptyprocess
pure-eval
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil
pytorch-warmup==0.1.1
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
responses==0.18.0
rich==13.3.2
rsa==4.9
safetensors==0.3.3
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.13.0
sentry-sdk==1.17.0
setproctitle==1.3.2
six
smmap==5.0.0
soundfile==0.12.1
stack-data
sympy==1.12
tabulate==0.9.0
tbb==2021.10.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
tifffile==2023.9.18
timm==0.6.12
tokenizers==0.13.2
tomli==2.0.1
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchstat==0.0.7
torchsummary==1.5.1
torchvision==0.15.1+cu118
tornado
tqdm==4.65.0
traitlets
transformers==4.27.1
triton==2.0.0
typing_extensions==4.4.0
urllib3==2.0.5
visualizer==0.0.1
wandb==0.14.0
wcwidth==0.2.6
Werkzeug==2.3.7
xxhash==3.2.0
yapf==0.32.0
yarl==1.8.2
zipp==3.15.0
